+++ include/async_runtime/fundamental/async/stream_subscription.h
@@ -6,7 +6,7 @@ template <>
 class StreamSubscription<std::nullptr_t> : public Object
 {
 public:
-    virtual void unsubscribe() = 0;
+    virtual void cancel() = 0;
 };
 
 template <typename T>
@@ -14,7 +14,7 @@ class StreamSubscription : public StreamSubscription<std::nullptr_t>
 {
 public:
     StreamSubscription(Function<void()> unsubscribe) : _unsubscribe(unsubscribe) {}
-    void unsubscribe() override
+    void cancel() override
     {
         _unsubscribe();
         _unsubscribe = [] {};