+++ include/async_runtime/widgets/stream_builder.h
@@ -8,7 +8,7 @@ class StreamBuilder : public StatefulWidget
     class _State;
 
 public:
-    using Builder = Funtion<ref<Widget>(ref<BuildContext>, ref<AsyncSnapshot<T>>)>
+    using Builder = Function<ref<Widget>(ref<BuildContext>, ref<AsyncSnapshot<T>>)>
     StreamBuilder(ref<Stream<T>> stream, Builder builder, option<Key> key = nullptr)
         : StatefulWidget(key),
           stream(stream), builder(builder) {}